# openapi.yaml
openapi: 3.0.3
info:
  title: CSM Events API
  version: 1.0.1
  description: Accepts batch CSM events and returns a correlation id.
servers:
  - url: /api/v1
paths:
  /csm-events:
    post:
      operationId: createCsmEvents
      summary: Create CSM events
      tags: [CSM Events]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FormCSMEventsRequest'
            examples:
              sample:
                value:
                  formCSMEvents:
                    - header:
                        eventStatus: abc
                        deliveryChannelId: "1234"
                      eventData:
                        eventDetailData:
                          svpSessionId: "4545"
                          formId: "7878"
                        employees:
                          - empId: "1566"
                            role: "Engineer"
                            department: "R&D"
                            tags: ["oncall","us"]
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateResponse'
        '400':
          description: Bad request
components:
  schemas:
    FormCSMEventsRequest:
      type: object
      required: [formCSMEvents]
      properties:
        formCSMEvents:
          type: array
          items:
            $ref: '#/components/schemas/FormCSMEvent'

    FormCSMEvent:
      type: object
      required: [header, eventData]
      properties:
        header:
          $ref: '#/components/schemas/Header'
        eventData:
          $ref: '#/components/schemas/EventData'

    Header:
      type: object
      required: [eventStatus, deliveryChannelId]
      properties:
        eventStatus: { type: string }
        deliveryChannelId: { type: string }

    EventData:
      type: object
      required: [eventDetailData, employees]
      properties:
        eventDetailData:
          $ref: '#/components/schemas/EventDetailData'
        employees:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/Employee'

    EventDetailData:
      type: object
      required: [svpSessionId, formId]
      properties:
        svpSessionId: { type: string }
        formId: { type: string }

    Employee:
      type: object
      required: [empId]
      properties:
        empId: { type: string }
      additionalProperties: true   # allow dynamic fields

    CreateResponse:
      type: object
      properties:
        status: { type: string, example: accepted }
        correlationId: { type: string, format: uuid }